The Minimal Homework
Created a user model
Created an AccountRepository interface with the methods save and login to register and login a user
Added the AccountRepositoryJDBC which implements the interface and apply the methods' logic!
Added the SaveServlet for register, MinimalServlet for the minimal.html page, and the UserLogin in for the userlogin.html page

Notes to take :
In case of a successful registration, the user will be redirected to the minimal.html page
In case of a successfull login, it will be redirected to a greeting response with their email, else the login form will be refreshed!